updated review

+++ PaperCastle.Core/Review.cs
@@ -11,7 +11,7 @@ namespace PaperCastle.Core
         public int Id { get; set; }
         public string Text { get; set; }
         public Book Book { get; set; }
-        public decimal Rating { get; set; }
+        public int Rating { get; set; }
         public ApplicationUser ApplicationUser { get; set; }
 
     }
